error if already exists"

+++ core/tools/implementations/createNewFile.ts
@@ -1,6 +1,7 @@
 import { inferResolvedUriFromRelativePath } from "../../util/ideUtils";
 
 import { ToolImpl } from ".";
+import { getCleanUriPath, getUriPathBasename } from "../../util/uri";
 
 export const createNewFileImpl: ToolImpl = async (args, extras) => {
   const resolvedFileUri = await inferResolvedUriFromRelativePath(
@@ -8,8 +9,26 @@ export const createNewFileImpl: ToolImpl = async (args, extras) => {
     extras.ide,
   );
   if (resolvedFileUri) {
+    const exists = await extras.ide.fileExists(resolvedFileUri);
+    if (exists) {
+      throw new Error(
+        `File ${args.filepath} already exists. Use the edit too to edit this file`,
+      );
+    }
     await extras.ide.writeFile(resolvedFileUri, args.contents);
     await extras.ide.openFile(resolvedFileUri);
+    return [
+      {
+        name: getUriPathBasename(resolvedFileUri),
+        description: getCleanUriPath(resolvedFileUri),
+        content: "File created successfuly",
+        uri: {
+          type: "file",
+          value: resolvedFileUri,
+        },
+      },
+    ];
+  } else {
+    throw new Error("Failed to resolve path");
   }
-  return [];
 };
